Fix substitution bug in genai/judges/utils.py
The format_prompt utility is being used for example to put the retrieval context into the prompt for the judge. It was using re.sub with the value as the second argument. In this form, re.sub handles some escape sequences like \1 to refer to the match that was found. If the retrieval context contains strings like `\1`, that leads to unintended substitutions, and if it contains an invalid escape sequence like `\x` it leads to an evaluation error due to an exception being thrown.

Fix this by using a lambda as a second argument to re.sub, resulting in the string being treated as plain text.

Signed-off-by: Stefan Walk <swalk@ptc.com>